Resolved Error Preventing Page components Loading
Previously the application was loaded but no components were loaded on the page and the console casted errors tha are complaining abou the pathname ie. Uncaught TypeError: Cannot read properties of  undefined (reading 'pathname') and some <Link> in aboutComponent. To which based on the resolution : https://stackoverflow.com/a/50072179 this happens when a react <Link> hasnt supplied with a to='' property