Change of application folder name
The name of the base folder for storing application files was changed from 'app-modules' to 'app' where the base folder was designed to contain all the files that are neccessary for website presentation and the other one for webapp presentation hence two application set 1. Website 2. Webapp where all the companies or community in the app will benefit with digital reach of their contents through the website implimentation and webapp will provide the community with the access to various web application for various business and process automations